+++ app/src/main/java/com/example/todoAppJpc/ui/todo/TodoEntryScreen.kt
@@ -1,29 +1,19 @@
 package com.example.todoAppJpc.ui.todo
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.statusBarsPadding
-import androidx.compose.foundation.layout.width
 import androidx.compose.material3.BottomAppBar
 import androidx.compose.material3.Button
-import androidx.compose.material3.DatePicker
-import androidx.compose.material3.DatePickerDialog
-import androidx.compose.material3.DatePickerState
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.material3.TextField
 import androidx.compose.material3.TimePicker
 import androidx.compose.material3.rememberDatePickerState
@@ -34,10 +24,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Dialog
-import androidx.compose.ui.window.DialogProperties
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.todoAppJpc.R
+import com.example.todoAppJpc.ui.components.DatePickerComponent
+import com.example.todoAppJpc.ui.components.TimePickerDialog
 import java.time.Instant
 import java.util.Calendar
 
@@ -165,113 +155,3 @@ fun TodoInputForm(
     }
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun DatePickerComponent(
-    viewModel: TodoEntryViewModel
-) {
-    val datePickerState = rememberDatePickerState(
-        initialSelectedDateMillis = Instant.now().toEpochMilli()
-    )
-    if (viewModel.getShowDatePicker()) {
-        Material3DatePickerDialogComponent(
-            datePickerState = datePickerState,
-            closePicker = {
-                viewModel.setShowDatePicker(false)
-                viewModel.setShowTimePicker(true)
-            },
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun Material3DatePickerDialogComponent(
-    datePickerState: DatePickerState,
-    closePicker: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    DatePickerDialog(
-        onDismissRequest = {
-            closePicker()
-        },
-        confirmButton = {
-            TextButton(
-                onClick = {
-                    datePickerState.setSelection(datePickerState.selectedDateMillis)
-                    closePicker()
-                }
-            ) {
-                Text(text = "OK")
-            }
-        },
-        dismissButton = {
-            TextButton(
-                onClick = {
-                    closePicker()
-                }
-            ) {
-                Text(text = "CANCEL")
-            }
-        },
-        modifier = modifier,
-    ) {
-        DatePicker(state = datePickerState)
-    }
-}
-
-@Composable
-fun TimePickerDialog(
-    title: String = "Select Time",
-    onCancel: () -> Unit,
-    onConfirm: () -> Unit,
-    toggle: @Composable () -> Unit = {},
-    content: @Composable () -> Unit,
-) {
-    Dialog(
-        onDismissRequest = onCancel,
-        properties = DialogProperties(
-            usePlatformDefaultWidth = false
-        ),
-    ) {
-        Surface(
-            shape = MaterialTheme.shapes.extraLarge,
-            tonalElevation = 6.dp,
-            modifier = Modifier
-                .width(IntrinsicSize.Min)
-                .height(IntrinsicSize.Min)
-                .background(
-                    shape = MaterialTheme.shapes.extraLarge,
-                    color = MaterialTheme.colorScheme.surface
-                ),
-        ) {
-            toggle()
-            Column(
-                modifier = Modifier.padding(24.dp),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Text(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(bottom = 20.dp),
-                    text = title,
-                    style = MaterialTheme.typography.labelMedium
-                )
-                content()
-                Row(
-                    modifier = Modifier
-                        .height(40.dp)
-                        .fillMaxWidth()
-                ) {
-                    Spacer(modifier = Modifier.weight(1f))
-                    TextButton(
-                        onClick = onCancel
-                    ) { Text("Cancel") }
-                    TextButton(
-                        onClick = onConfirm
-                    ) { Text("OK") }
-                }
-            }
-        }
-    }
-}
